upload with s==0 prtection

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: arenilla <arenilla@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/10 13:56:23 by arenilla          #+#    #+#             */
-/*   Updated: 2024/02/02 10:54:25 by arenilla         ###   ########.fr       */
+/*   Updated: 2024/02/15 08:23:24 by arenilla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
